Controller fiche ia          reste test & remplir modal Changement couleurs fiche tra & ia statut

+++ src/main/java/phenotypage/model/fiche/ficheIa/FicheIaServiceImpl.java
@@ -14,6 +14,7 @@ import phenotypage.model.vache.Vache;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 /**
@@ -47,6 +48,11 @@ FicheIaServiceImpl implements FicheIaService
 	}
 
 	@Override
+	public FicheIa findTopByOrderByNomDesc() {
+		return repository.findTopByOrderByNomDesc();
+	}
+
+	@Override
 	public FicheIa save(FicheIa ficheIa) {
 		return repository.save(ficheIa);
 	}
@@ -58,11 +64,23 @@ FicheIaServiceImpl implements FicheIaService
 
 	@Override
 	public FicheIa createFicheIa(String nom, Date dateFiche, String lieu, Programme programme, String numIpe, String numSemence, Vache vache, Insemination insemination, Traitement_Donneuse traitement_donneuse, Gestation gestation) {
-		Traitement_Donneuse traitement = traitement_donneuseService.createTraitement_Donneuse(traitement_donneuse);
-		Insemination inseminationSave = inseminationService.createInsemination(insemination);
-		Gestation gestationSave = gestationService.createGestation(gestation);
 
 		FicheIa ficheIa = new FicheIa();
+
+		if(insemination != null) {
+			Insemination inseminationSave = inseminationService.createInsemination(insemination);
+			ficheIa.setInsemination(inseminationSave);
+		}
+
+		if(traitement_donneuse != null) {
+			Traitement_Donneuse traitement = traitement_donneuseService.createTraitement_Donneuse(traitement_donneuse);
+			ficheIa.setTraitement_donneuse(traitement);
+		}
+		if(gestation != null) {
+			Gestation gestationSave = gestationService.createGestation(gestation);
+			ficheIa.setGestation(gestationSave);
+		}
+
 		ficheIa.setNom(nom);
 		ficheIa.setDateHeureMinute(dateFiche);
 		ficheIa.setLieu(lieu);
@@ -70,38 +88,83 @@ FicheIaServiceImpl implements FicheIaService
 		ficheIa.setNumIpe(numIpe);
 		ficheIa.setNumDepotSemence(numSemence);
 		ficheIa.setVache(vache);
-		ficheIa.setInsemination(inseminationSave);
-		ficheIa.setTraitement_donneuse(traitement);
-		ficheIa.setGestation(gestationSave);
+
+		ficheIa.setStatut(determineStatut(ficheIa));
 
 		return save(ficheIa);
 	}
 
 	@Override
 	public FicheIa updateFicheIa(FicheIa ficheIaForUpdate, String nom, Date dateFiche, String lieu, Programme programme, String numIpe, String numSemence, Vache vache, Insemination insemination, Traitement_Donneuse traitement_donneuse, Gestation gestation) {
-		Traitement_Donneuse traitementDonneuseToDelete = ficheIaForUpdate.getTraitement_donneuse();
-		Gestation gestationToDelete = ficheIaForUpdate.getGestation();
+		if(traitement_donneuse != null){
+			if(ficheIaForUpdate.getTraitement_donneuse() != null) {
+				if (!Objects.equals(traitement_donneuse.getId(), ficheIaForUpdate.getTraitement_donneuse().getId())) {
+					Traitement_Donneuse traitementDonneuseToDelete = ficheIaForUpdate.getTraitement_donneuse();
+					ficheIaForUpdate.setTraitement_donneuse(traitement_donneuseService.createTraitement_Donneuse(traitement_donneuse));
+					traitement_donneuseService.delete(traitementDonneuseToDelete);
+				}
+			}else{
+				ficheIaForUpdate.setTraitement_donneuse(traitement_donneuseService.createTraitement_Donneuse(traitement_donneuse));
+			}
+		}
+
+		if(gestation != null){
+			if(ficheIaForUpdate.getGestation() != null) {
+				if (!Objects.equals(gestation.getId(), ficheIaForUpdate.getGestation().getId())) {
+					Gestation gestationToDelete = ficheIaForUpdate.getGestation();
+					ficheIaForUpdate.setGestation(gestationService.createGestation(gestation));
+					gestationService.delete(gestationToDelete);
+				}
+			}else{
+				ficheIaForUpdate.setGestation(gestationService.createGestation(gestation));
+			}
+		}
+
+		if(insemination != null){
+			if(ficheIaForUpdate.getInsemination() != null) {
+				if (!Objects.equals(insemination.getId(), ficheIaForUpdate.getGestation().getId())) {
+					Insemination inseminationToDelete = ficheIaForUpdate.getInsemination();
+					ficheIaForUpdate.setInsemination(inseminationService.createInsemination(insemination));
+					inseminationService.delete(inseminationToDelete);
+				}
+			}else{
+				ficheIaForUpdate.setInsemination(inseminationService.createInsemination(insemination));
+			}
+		}
+
 
 		ficheIaForUpdate.setNom(nom);
-		ficheIaForUpdate.setNom(nom);
-		ficheIaForUpdate.setDateHeureMinute(dateFiche);
-		ficheIaForUpdate.setLieu(lieu);
 		ficheIaForUpdate.setProgramme(programme);
+		ficheIaForUpdate.setDateHeureMinute(dateFiche);
 		ficheIaForUpdate.setNumIpe(numIpe);
 		ficheIaForUpdate.setNumDepotSemence(numSemence);
+		ficheIaForUpdate.setLieu(lieu);
 		ficheIaForUpdate.setVache(vache);
-		ficheIaForUpdate.setInsemination(inseminationService.updateInsemination(insemination));
-		ficheIaForUpdate.setTraitement_donneuse(traitement_donneuseService.createTraitement_Donneuse(traitement_donneuse));
-		ficheIaForUpdate.setGestation(gestationService.createGestation(gestation));
 
-		gestationService.delete(gestationToDelete);
-		traitement_donneuseService.delete(traitementDonneuseToDelete);
+		ficheIaForUpdate.setStatut(determineStatut(ficheIaForUpdate));
 
 		return save(ficheIaForUpdate);
 	}
 
 	@Override
-	public FicheIa findTopByOrderByNomDesc() {
-		return repository.findTopByOrderByNomDesc();
+	public int determineStatut(FicheIa ficheIa) {
+		if(Objects.equals(ficheIa.getNom(), "") || ficheIa.getDateHeureMinute() == null || ficheIa.getVache() == null){
+			return 2;
+		}else if(ficheIa.getProgramme() == null || Objects.equals(ficheIa.getNumIpe(), "") || Objects.equals(ficheIa.getLieu(), "")
+				|| ficheIa.getOperateur() == null || ficheIa.getTraitement_donneuse() == null || ficheIa.getInsemination() == null
+				|| Objects.equals(ficheIa.getNumDepotSemence(), "") || ficheIa.getGestation() == null){
+			return 1;
+		}else{
+			if(inseminationService.determineStatut(ficheIa.getInsemination()) == 1){
+				return 1;
+			}
+			if(traitement_donneuseService.determineStatut(ficheIa.getTraitement_donneuse()) == 1){
+				return 1;
+			}
+			if(gestationService.determineStatut(ficheIa.getGestation()) == 0){
+				return 1;
+			}
+		}
+		return 0;
 	}
 }
